Wire up EditFunnel analytics for image tags.

+++ app/src/main/java/org/wikipedia/analytics/EditFunnel.java
@@ -34,7 +34,7 @@ public class EditFunnel extends Funnel {
         );
     }
 
-    public void logSaved(int revID) {
+    public void logSaved(long revID) {
         log(
                 "action", "saved",
                 "revID", revID
